stable version but admin buttons do not work
