+++ konnekt_app/app/[id].tsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from 'react';
-import { View, Text, StyleSheet, TouchableOpacity, Alert, ActivityIndicator } from 'react-native';
-import { useLocalSearchParams, useRouter } from 'expo-router';
+import { View, Text, StyleSheet, TouchableOpacity, Alert, ActivityIndicator, Switch } from 'react-native';
+import { useLocalSearchParams } from 'expo-router';
 import * as Location from 'expo-location';
+import { IP_ADDRESS } from '../src/components/config/globalvariables';
+import useAuthRedirect from '../src/hooks/useAuthRedirect';
 
 import ProximityChecker from '../src/components/ProximityChecker';
 import ClubMembersPanel from '../src/components/pages/ClubMembersPanel';
-import { IP_ADDRESS } from '../src/components/config/globalvariables';
-import useAuthRedirect from '../src/hooks/useAuthRedirect';
 
 type Club = {
   _id: string;
@@ -21,43 +21,45 @@ type Club = {
     latitude: number;
     longitude: number;
   };
+  joinCode: string;
+  pending: User[];
+  members: User[];
+};
+
+type User = {
+  _id: string;
+  username: string;
+  full_name: string;
+  meetingsAttended?: number;
 };
 
 export default function ClubDetailScreen() {
   useAuthRedirect();
   const { id } = useLocalSearchParams();
-  const router = useRouter();
-
   const [club, setClub] = useState<Club | null>(null);
   const [location, setLocation] = useState<Location.LocationObject | null>(null);
   const [loading, setLoading] = useState(true);
+  const username = global.authUser?.username ?? '';
 
-  useEffect(() => {
-    if (!id || typeof id !== 'string') {
-      Alert.alert("Error", "Invalid club ID.");
-      return;
-    }
-
-    const fetchClub = async () => {
-      try {
-        const response = await fetch(`http://${IP_ADDRESS}:5000/api/clubs/${id}`);
-        const data = await response.json();
-        if (response.ok) {
-          setClub(data);
-        } else {
-          Alert.alert("Error", data.error || "Club not found.");
-        }
-      } catch (err) {
-        console.error("❌ Failed to fetch club:", err);
-        Alert.alert("Error", "Could not load club info.");
-      } finally {
-        setLoading(false);
+  // Fetch club details
+  const fetchClub = async () => {
+    try {
+      const response = await fetch(`http://${IP_ADDRESS}:5000/api/clubs/${id}`);
+      const data = await response.json();
+      if (response.ok) {
+        setClub(data);
+      } else {
+        Alert.alert("Error", data.error || "Club not found.");
       }
-    };
+    } catch (err) {
+      console.error("❌ Failed to fetch club:", err);
+      Alert.alert("Error", "Could not load club info.");
+    } finally {
+      setLoading(false);
+    }
+  };
 
-    fetchClub();
-  }, [id]);
-
+  // Handle location tracking
   const handleCheckIn = async () => {
     setLoading(true);
 
@@ -79,6 +81,7 @@ export default function ClubDetailScreen() {
     }
   };
 
+  // Set club check-in location
   const handleSetLocation = async () => {
     try {
       const { status } = await Location.requestForegroundPermissionsAsync();
@@ -110,12 +113,73 @@ export default function ClubDetailScreen() {
     }
   };
 
+  // Update club settings (like public, location tracking)
+  const handleUpdateToggle = async (updates: Partial<Club>) => {
+    if (!club) return;
+    try {
+      const res = await fetch(`http://${IP_ADDRESS}:5000/api/clubs/${club._id}`, {
+        method: 'PATCH',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ updates })
+      });
+      if (res.ok) setClub(prev => prev ? { ...prev, ...updates } : prev);
+    } catch (err) {
+      Alert.alert("Error", "Failed to update");
+    }
+  };
+
+  // Promote or demote members
+  const handlePromoteDemote = async (userId: string, promote: boolean) => {
+    if (!club) return;
+    const endpoint = promote ? 'promote' : 'demote';
+    await fetch(`http://${IP_ADDRESS}:5000/api/clubs/${club._id}/${endpoint}/${userId}`, {
+      method: 'PATCH'
+    });
+    fetchClub();
+  };
+
+  // Kick members from club
+  const handleKick = async (userId: string) => {
+    if (!club) return;
+    await fetch(`http://${IP_ADDRESS}:5000/api/clubs/${club._id}/kick/${userId}`, {
+      method: 'PATCH'
+    });
+    fetchClub();
+  };
+
+  // Handle approval or rejection of pending members
+  const handleApproval = async (userId: string, approve: boolean) => {
+    if (!club) return;
+    const endpoint = approve ? 'approve' : 'reject';
+    try {
+      const res = await fetch(`http://${IP_ADDRESS}:5000/api/clubs/${club._id}/${endpoint}/${userId}`, {
+        method: 'PATCH',
+      });
+  
+      if (res.ok) {
+        Alert.alert("Success", `User ${approve ? 'approved' : 'rejected'}`);
+        fetchClub(); // Reload club details
+      } else {
+        const data = await res.json();
+        Alert.alert("Error", data.error || "Failed to update user status.");
+      }
+    } catch (err) {
+      console.error("❌ Approval error:", err);
+      Alert.alert("Error", "Could not update user status.");
+    }
+  };
+
+  useEffect(() => {
+    if (!id || typeof id !== 'string') {
+      Alert.alert("Error", "Invalid club ID.");
+      return;
+    }
+    fetchClub();
+  }, [id]);
+
   if (loading || !club) {
     return <ActivityIndicator size="large" color="#4c87df" style={{ marginTop: 50 }} />;
   }
-
-  const isAdmin = club.admins.includes(global.authUser?.username ?? '');
-  const isOwner = club.owner === global.authUser?.username;
 
   return (
     <View style={styles.container}>
@@ -125,15 +189,6 @@ export default function ClubDetailScreen() {
       <Text style={styles.statusText}>
         📍 Location Check-In is {club.useLocationTracking ? 'enabled ✅' : 'disabled ❌'}
       </Text>
-
-      {(isAdmin || isOwner) && (
-        <TouchableOpacity
-          style={styles.setLocButton}
-          onPress={() => router.push({ pathname: '/admin-panel', params: { id: club._id } })}
-        >
-          <Text style={styles.buttonText}>Admin Settings</Text>
-        </TouchableOpacity>
-      )}
 
       <TouchableOpacity style={styles.button} onPress={handleCheckIn}>
         <Text style={styles.buttonText}>
@@ -145,12 +200,52 @@ export default function ClubDetailScreen() {
         <ProximityChecker anchor={club.checkInCoords} />
       )}
 
-      <ClubMembersPanel
-        clubId={club._id}
-        currentUserId={global.authUser?.username ?? ''}
-        isAdmin={isAdmin}
-        isOwner={isOwner}
-      />
+      <Text style={styles.subheader}>Admin Options</Text>
+      <View style={styles.toggleRow}>
+        <Text>Public Club</Text>
+        <Switch
+          value={club.isPublic}
+          onValueChange={(val) => handleUpdateToggle({ isPublic: val })}
+        />
+      </View>
+
+      <View style={styles.toggleRow}>
+        <Text>Location Tracking</Text>
+        <Switch
+          value={club.useLocationTracking}
+          onValueChange={(val) => handleUpdateToggle({ useLocationTracking: val })}
+        />
+      </View>
+
+      <TouchableOpacity style={styles.button} onPress={handleSetLocation}>
+        <Text style={styles.buttonText}>Set Check-in Location</Text>
+      </TouchableOpacity>
+
+      <Text style={styles.subheader}>Join Code: {club.joinCode}</Text>
+
+      <Text style={styles.subheader}>Pending Approvals</Text>
+      {club.pending?.map(user => (
+        <View key={user._id} style={styles.userRow}>
+          <Text>{user.full_name}</Text>
+          <TouchableOpacity onPress={() => handleApproval(user._id, true)}><Text>✅</Text></TouchableOpacity>
+          <TouchableOpacity onPress={() => handleApproval(user._id, false)}><Text>❌</Text></TouchableOpacity>
+        </View>
+      ))}
+
+      <Text style={styles.subheader}>Members</Text>
+      {club.members?.map(user => (
+        <View key={user._id} style={styles.userRow}>
+          <Text>{user.full_name}</Text>
+          {club.admins.includes(user.username) ? (
+            <TouchableOpacity onPress={() => handlePromoteDemote(user._id, false)}><Text>⬇️</Text></TouchableOpacity>
+          ) : (
+            <TouchableOpacity onPress={() => handlePromoteDemote(user._id, true)}><Text>⬆️</Text></TouchableOpacity>
+          )}
+          {user._id !== username && (
+            <TouchableOpacity onPress={() => handleKick(user._id)}><Text>🗑️</Text></TouchableOpacity>
+          )}
+        </View>
+      ))}
     </View>
   );
 }
@@ -181,13 +276,6 @@ const styles = StyleSheet.create({
     borderRadius: 8,
     marginTop: 20,
   },
-  setLocButton: {
-    backgroundColor: '#2e7d32',
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-    borderRadius: 8,
-    marginBottom: 10,
-  },
   buttonText: {
     color: '#fff',
     fontWeight: '600',
@@ -198,5 +286,23 @@ const styles = StyleSheet.create({
     fontWeight: 'bold',
     color: '#333',
     marginBottom: 10,
+  },
+  subheader: {
+    fontSize: 18,
+    fontWeight: '600',
+    marginTop: 20,
+    marginBottom: 8,
+  },
+  userRow: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    paddingVertical: 6
+  },
+  toggleRow: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginVertical: 10
   },
 });
